improved file structure, and norminated
